requirements.txt
python-barcode
qrcode
Pillow
opencv-python
pyzbar